+++ src/modules/InvitationForPunchOut/views/SearchIPO/Table/InvitationsTable.style.tsx
